requirements.txt
python-dotenv
requests
flask
pymongo
qrcode
pillow